pyproject.toml
[project]
name = "rag-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "instructorembedding>=1.0.1",
    "matplotlib>=3.10.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.4.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
]
